+++ src/doc/changelog.md
@@ -5,6 +5,7 @@
 - Add overview map feature
 - Add scale line feature
 - Add swipe feature
+- Add heatmap layer
 
 NOTE: This WireCloud widget is based on Wirecloud/ol3-map-widget created by
 CoNWeT Lab., Universidad Politecnica de Madrid and Future Internet Consulting